+++ package.json
@@ -32,7 +32,6 @@
     "container-query-polyfill": "^1.0.2",
     "date-fns": "^2.29.3",
     "eslint-config-prettier": "^8.8.0",
-    "filbert": "^0.1.20",
     "file-saver": "^2.0.5",
     "fs-extra": "^9.0.1",
     "graphql": "^16.6.0",
