+++ src/components/Editor/Runners/PyodideRunner/PyodideRunner.jsx
@@ -22,7 +22,7 @@ import RunnerControls from "../../../RunButton/RunnerControls";
 import PyodideWorker from "worker-loader!./PyodideWorker.js";
 import { switchToSkulpt } from "../../../../redux/RunnerSlice";
 
-const PyodideRunner = () => {
+const PyodideRunner = ({ active }) => {
   const pyodideWorker = useMemo(() => new PyodideWorker(), []);
   const interruptBuffer = useRef();
   const stdinBuffer = useRef();
@@ -46,6 +46,7 @@ const PyodideRunner = () => {
   const showVisualTab = queryParams.get("show_visual_tab") === "true";
   const [hasVisual, setHasVisual] = useState(showVisualTab || senseHatAlways);
   const [visuals, setVisuals] = useState([]);
+  const [showRunner, setShowRunner] = useState(true);
 
   useEffect(() => {
     pyodideWorker.onmessage = ({ data }) => {
@@ -82,13 +83,19 @@ const PyodideRunner = () => {
   }, []);
 
   useEffect(() => {
-    if (codeRunTriggered) {
+    if (codeRunTriggered && active) {
       handleRun();
     }
   }, [codeRunTriggered]);
 
   useEffect(() => {
-    if (codeRunStopped) {
+    if (codeRunTriggered) {
+      setShowRunner(active);
+    }
+  }, [codeRunTriggered]);
+
+  useEffect(() => {
+    if (codeRunStopped && active) {
       handleStop();
     }
   }, [codeRunStopped]);
@@ -277,7 +284,10 @@ const PyodideRunner = () => {
   };
 
   return (
-    <div className={`pythonrunner-container`}>
+    <div
+      className={`pythonrunner-container pyodiderunner`}
+      style={{ display: showRunner ? "block" : "none" }}
+    >
       {isSplitView ? (
         <>
           {hasVisual && (
